% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionGraph.R
\name{DecisionGraph}
\alias{DecisionGraph}
\title{Decision Graph}
\usage{
DecisionGraph(W, I, Density, Log = "xy")
}
\arguments{
\item{W}{edge weight vector}

\item{I}{parent node vector}

\item{Density}{density vector}
}
\value{
peaks: indexes of the densit-peak nodes
}
\description{
Decision Graph for determing the inter-cluster edge nodes
}
