% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate_clustering_result.R
\name{Evaluate_clustering_result}
\alias{Evaluate_clustering_result}
\title{Evaluate clustering results}
\usage{
Evaluate_clustering_result(data_name, Label, annotation_data,
  EvaIndex = c("NMI", "ARI", "MeanFscore"))
}
\arguments{
\item{Label, annotation_data, EvaIndex}{}

\item{EvaIndex}{}
}
\description{
Evaluate clustering results
my_evaluate_clustering_result
}
\examples{

}
