% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HNND.R
\name{HNND}
\alias{HNND}
\title{Hierachical Nearest Neighbor Descent}
\usage{
HNND(x, K = 20, h = 50, disName = "euclidean",
  reuse_tree_contruction = "off")
}
\description{
Hierachical Nearest Neighbor Descent (equivalent to Hierachical Nearest Neighbor Ascent)
}
\examples{

}
\keyword{HNND}
