% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HNND_C.R
\name{HNND_C}
\alias{HNND_C}
\title{Clustering by Hierachical Nearest Neighbor Descent}
\usage{
HNND_C(x, K = 20, h = 20, disName = "euclidean", LogPlot = "")
}
\arguments{
\item{x}{dataset (a matrix with rows denoting samples and columns denoting features}

\item{K}{number of nearest neighbors (emperical value: the nearest integer to log2(N), where N is the size of the dataset)}

\item{h}{a threshold for deciding when to start the ND layer (usually h = k)}

\item{disName}{distance measurement ('euclidean','cosine', and'l2' are supported)}

\item{LogPlot}{whether to show the decision graph in the log scale}
}
\value{
I:    parent node vector;

W:    edge weigth vector;

Label:    cluster label vector;

Totoltime_DG:     time spent on the interactive operation on the Decision Graph (which is user-dependent and is thus suggested to be subtracted from the total runtime);
}
\description{
Clustering by Hierachical Nearest Neighbor Descent (equivalent to: Clustering by Hierachical Nearest Neighbor Ascent)
}
\examples{
data('cytof.benchmark.h1')
x = cytof.benchmark.h1$x
K = ceiling(log2(nrow(x)))
result <- HNND_C(x = x,K = K,h = K,disName = "euclidean",LogPlot = '')

}
\author{
Teng Qiu
}
