% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ND.R
\name{ND}
\alias{ND}
\title{Nearest Descent}
\usage{
ND(F, x, disName = "euclidean")
}
\arguments{
\item{F}{density vector}

\item{x}{test dataset}

\item{disName}{distanc measurement}
}
\value{
W:   edge weight vector;

I:   parent node vector;

F:   density vector;

root:   index of the root node (there is only one root node);
}
\description{
Nearest Descent (equivalent to Nearest Ascent)
}
\author{
Teng Qiu
}
