% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NND.R
\name{NND}
\alias{NND}
\title{Nearest Neighbor Descent}
\usage{
NND(F, knn)
}
\arguments{
\item{F}{density vector}

\item{knn}{a list}
}
\value{
W:   edge weight vector;

I:   parent node vector;

roots:   indexes of the root nodes;

K_optimal:   which neighbor each parent node is;
}
\description{
Nearest Neighbor Descent (equivalent to Nearest Neighbor Ascent)
}
\author{
Teng Qiu
}
