% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryKnnHNSW.R
\name{queryKnnHNSW}
\alias{queryKnnHNSW}
\title{Find k-nearest-neighbors (by HNSW)}
\usage{
queryKnnHNSW(ann, x, K, disName)
}
\arguments{
\item{ann}{constructed search index}

\item{x}{dataset}

\item{K}{number of nearest neighbors}

\item{disName}{distanc measurement}
}
\value{
nn.index:   a neighbor index matrix

nn.dist:   a neighbor distance matrix
}
\description{
get the k-nearest-neighbors of the query points
}
