% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simlified_knn_density_estimation_v1.R
\name{simlified_knn_density_estimation_v1}
\alias{simlified_knn_density_estimation_v1}
\title{density estimation}
\usage{
simlified_knn_density_estimation_v1(knn)
}
\arguments{
\item{knn}{a list (knn$nn.dist: knn distance matrix)}
}
\value{
F: density vector
}
\description{
simlified knn density estimation (v1)
}
