% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DecisionGraph.R
\name{DecisionGraph}
\alias{DecisionGraph}
\title{Decision Graph}
\usage{
DecisionGraph(result, Log = "xy")
}
\description{
Decision Graph for determing the inter-cluster edge nodes
}
\examples{

}
\keyword{DecisionGraph}
